Move implicits functionality to the implicits module
